Rename test/unit/capabilities.js to test/unit/airtap.js
